+++ Cargo.toml
@@ -20,7 +20,7 @@ clap = {version = "4.5.20", features=["derive"]}
 docx-rs = {version="0.4.17", optional=true}
 log = "0.4.22"
 once_cell = "1.20.2"
-pyo3 = { version = "0.23.3", features = ["extension-module"] }
+pyo3 = { version = "0.23.3", features = ["extension-module"], optional = true}
 regex = "1.11.1"
 serde = {version = "1.0.210", features=["derive"]}
 serde_json = "1.0.128"
@@ -35,3 +35,4 @@ tempfile = "3.14.0"
 
 [features]
 docx = ["dep:docx-rs"]
+python = ["dep:pyo3"]
